Fix errors generated from PR #307 merge (#311)
* Fix - direct messaging '404' error (#109, #307)

PR #307 introduced direct messaging functionality. This works fine on top-level pages, but due to missing slashes in the `users` and `direct-messages` requests a 404 error is generated when on a sub-level page (e.g. `/projects/`).

* Updated `taskBoard.vue` to use updated API response

PR #307 updated the `/users` API to use "users" rather than "data" in the JSON response generated in `app/Http/Controllers/UserController.php`.

This commit stops an error being thrown ("`Uncaught (in promise) TypeError: Cannot read property 'data' of undefined`") when trying to open a task.